Append launch parameters to the end of the arguments list
The wrapper keeps track of a bunch of launch parameters for itself and those to be passed to a client/server, however, the recent update to how arguments are handled did not include the necessary for loop to append said parameters. This meant that arguments set by a launcher/user are parsed as normal, but the wrapper never appended its parameters to the end of the arguments list. This broke versions which expected parameters such as 'mppass' to always be present.